+++ include/ShaderWriter/FragmentWriter.inl
@@ -202,7 +202,7 @@ namespace sdw
 	inline T FragmentWriter::declBlendOutput( std::string const & name
 		, uint32_t location
 		, uint32_t blendIndex
-		, uint32_t attributes
+		, uint64_t attributes
 		, bool enabled )
 	{
 		static_assert( !IsSameV< T, Boolean >, "Bool is not supported as output type" );
@@ -251,7 +251,7 @@ namespace sdw
 		, uint32_t location
 		, uint32_t blendIndex
 		, uint32_t dimension
-		, uint32_t attributes
+		, uint64_t attributes
 		, bool enabled )
 	{
 		static_assert( !IsSameV< T, Boolean >, "Bool is not supported as output type" );
